% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{alphainfo_plans}
\alias{alphainfo_plans}
\title{List available billing plans}
\usage{
alphainfo_plans(client)
}
\arguments{
\item{client}{An \code{alphainfo_client}.}
}
\value{
A named list of available billing plans.
}
\description{
List available billing plans
}
